working logs
