+++ examples/exampledata/config/ng_pipeline.yml
@@ -5,7 +5,7 @@ restart_count: 2
 config_refresh_interval: 5
 error_backlog_size: 1500000
 logger:
-  level: DEBUG
+  level: INFO
   format: "%(asctime)-15s %(hostname)-5s %(name)-10s %(levelname)-8s: %(message)s"
   datefmt: "%Y-%m-%d %H:%M:%S"
   loggers:
